Drop the "Search Again" menu item that appears after searching
When you have done a search a "Search Again" item appears in the nav
menu that will take you back to the the page you searched (simple or
advanced query).

But this seems to be completely unnecessary because the simple and
advanced query have their own menu items which the user would have
followed in the first place.

So let's get rid of this Search Again menu item.