Move the content parts of index to content/index-content.jsp
The goal of this commit is to enable non Java/JSP people to edit
the index page content without having to understand too much of
how FRED hangs together.

With this in mind the main content part of the page is moved
to `content/index-content.jsp` and is slurped into the index by
a jsp:include. This content would have become static html if it
wasn't for a couple of dynamic bits to show the record count and
update date.