Add a login form to index-content, but only if user not logged in
This commits adds a log in form to the index page that is only
rendered if the current user is not already logged in.

This form submits to the existing `login.jsp` which will be displayed
should anything go wrong. Like the user mis-enters their password.

`login-form.jspf` is a jsp fragment that contains the login form html.
It was copy pasted from how the existing login form is rendered then
I wired the recapthca token back into it. This was rather than
attempting to inherit from FREDIPSysLoginPage which seemed quite wrong.

The down side of the current approach though is that if the underlying
library changes how the login form is rendered it won't be updated here.
Let's assume that is unlikely to be a problem.