Download, speed up the file save
Currently when you go to download results you can be left waiting
for some time before the file download starts. This is because there
is a lot of work to fetch the things to download that is currently
been down before any results are written.

So let's shake things up a little and:
- write the file header straight away. The user will see some immediate
  results now.
- then we fetch the data that will be downloaded. so the speed will seem
  to tank then.
- drop the no results handling. now if there are no results you'll get a
  file download with no results (e.g header only)