detail page, render what you can fast
Currently for some features the display of the deails page
can take a long time to show anything.

Do a quick search for 'J36/f9524` for an example.

The page is getting bogged down in the taxa sections. But
there is a lot of useful content before this section that
could be displayed.

We can add in some calls to `out.flush()` to have whatever
content is available pushed out to the browser. The rest
of the page will still be loading but at least the user
has some info to read a lot sooner.