Fix the downloads for non locality options
The downloads were split into different parts then changed
so that the file header was written out before the hard work
of fetch all the samples to put into the download. This was
so the user doing the download would get a download prompt
straight away.

But when doing this that hard work of fetching the samples to
download was put into the locality only branch which resulted
in all the other downloads having no content. Because we hadn't
fetched the samples at all.

So we need to move this code to the common part.