Detail page, inhibit the NPC taxon links
The detail page lists every taxon found at the locality.

And currently for every taxon it does an HTTP request to NPC to
ensure it is known and render a link to it. Otherwise just the
name (without link) is rendered.

But this really slows the page load down. By quite a lot.

So for now lets inhibit the check. We would rather have a fast
page load and no links.