Pass the session into FREDIPSysJspPage.getFREDTemplate
We will need the session to work out if the current user is
logged in. This will be coming up in the next commit

So in this commit we pass the session in as an argument to
getFREDTemplate. And update all the places that call this
method to pass in that session.