FRFormServlet add a FeatureID parameters that will fetch all the things
Currently the detail page will build a frf link to print the whole
fossil record and that link will contain all the sample and record ids
found at the feature. Building this link has a huge up front cost as
the entire entity model must be traversed and it's just a link, so
most of the time it won't even be used.

For large drillholes like Maui this can even make the page fail to load.

So to resolve this we add a new FeatureID parameter to the FRF servlet
and have the servlet find the samples/records to print for the given
feature.