PDFUtil, don't be putting `null` in repeating cells
PDFUtil.addRepeatingCells is used to populate a block of repeating table
cells. The first column is expected to be a heading following by value
columns.

The heading is only printed for the first row. But on subsequent rows
`null` is printed instead. So we change that to an empty string.

If this method is passed an empty array it is currently printing
`null` if the value column. `null` is not a very friendly term unless
you're a programmer. So lets replace that with `N/A` which will
hopefully make more sense to general users.