FRFormServlet, don't print `null`
The PDF's generated by this servlet have `null` printed in a bunch of
places. The underlying pdf generator will print the string `null` for
a null value so we'd be better off mapping null to some place holder
string. Either "" (empty string) or "N/A".

And that is what is done in this commit.

There is also a bunch of repeating code of the form:
	`x != null ? x.method() : null`
I've added some utility methods that take care of this boiler plate
and hopefully make it easier to see what is happening in the code.