detail page, make useful PDF filenames
The detail page contains many links to get PDF documents. But currently
all of these links have the same meaningless filename `frf.pdf`.

The FrFormServlet (that generates the PDFs) doesn't even look at the
filename part of the request. But the browser will use it should the
PDF be saved.

So let's give the link a filename that will describe the file contents
and hopefully make FRED users happy.