Fix NPE on detail page for new localities
62ab40d21 changed the PDF links to contain the FR number. But this
change didn't consider the possiblity that the FR number may be null.
And of course the FR number will be null for localities that are in the
process of being added.

E.g. go to Datafolders, add new drillhole then view the new locality and
boom. Well not with this fix.