Make a SimpleSearchServlet to front simple search requests
Currently `simple_query.jsp` is building an HQL query on the page
in a form submit handler. This is something that is all sorts of wrong
and would be a security analyst field day.

So in this commit we:
- add a servlet to act as a controller. it will accept the search
  parameters and build the query. This controller is mapped to the
  existing `simple_query.jsp` path
- simple_query.jsp is:
  - moved to /WEB-INF so it won't be directly served. You need to go via the front controller.
  - posts the search form to the front controller (to build the query
    there)
- ResultList is updated to read the query from the request attributes
  populated by the search servlet. No more running user supplied queries

At this stage the page is still building the query but this is useful
to confirm the serlet is doing the right thing. I'll be going later.