Fix rendering of advanced and confidential search
Currently the advanced and confidential searches are rendered in a
frameset. This is not great because the html frameset is:
- deprecated (see
  https://developer.mozilla.org/en-US/docs/Web/HTML/Element/frameset)
- a little bit inside out logically adv_query.jsp/adv_query_record.jsp
  are a child of the search they are in. But in the frameset they are a
  sibling which is why they render over the top of bottom part of the
  nav menu.

The solution here is to switch to using an iframe. This means we can
drop the now unnecessary frameset although we do need to add some
complexity in using javascript for the child iframe's content to update
it's containers height.