Simple search, auto fix year and age ranges
Currently a bad year range (1990-1980) results in no data and a bad Age
range results in a javascript warning telling you to fix it up.

But the consensus ages will fix up whatever range you give it. E.g the
user just needs to enter endpoints and the code will figure out the
from/to from that. This is much nicer for our users.

So lets do this for year and age as well.