Add check for null feature site id
Do a null check on the feature siteId before attempting to get the site
details.

This can happen during feature creation if the supplied
coordinates were invalid. In that case the siteId for the feature will
be null.

A subsequent commit will improve the error checking but this commit
allows any existing features to be edited.

Signed-off-by: Lockyer <g.lockyer@gns.cri.nz>